download and plot memory usage

+++ metrics/downloadMetrics.R
@@ -14,7 +14,7 @@ date_ranges <- rbind(
   # data.frame(from = "2025-05-01 00:00:00 UTC", until = "2025-05-31 23:59:59 UTC")
 )
 
-start_date <- as.Date("2024-06-01")
+start_date <- as.Date("2025-01-01")
 
 while (start_date < Sys.Date()) {
   next_start_date <- as.Date(gsub("([[:digit:]]{4}-[[:digit:]]{2})-[[:digit:]]{2}", "\\1-01", as.character(start_date + 31)))
@@ -26,6 +26,7 @@ while (start_date < Sys.Date()) {
   date_to <- paste(end_date, "00:00:00 UTC")
   message(sprintf("{%s} - {%s}", date_from, date_to))
   year_month <- gsub(pattern = "([[:digit:]]{4}-[[:digit:]]{2}).+", replacement = "\\1", x = date_from)
+  # connection counts
   file_out <- paste0(year_month, "-connect_count.rds")
   if (!file.exists(file_out)) {
     try({
@@ -42,26 +43,22 @@ while (start_date < Sys.Date()) {
       saveRDS(metrics_df, file_out)
     })
   }
+  # memory usage
+  file_out <- paste0(year_month, "-total_rss")
+  if (!file.exists(file_out)) {
+    try({
+      metrics_df <- rsconnect::showMetrics(
+        metricSeries = "docker_container_mem",
+        metricNames = c("total_rss"),
+        appName = "ifnresource",
+        account = "rehwinkellab",
+        server="shinyapps.io",
+        from = as.numeric(as.POSIXct(date_from)),
+        until = as.numeric(as.POSIXct(date_to)),
+        interval = "60s"
+      )
+      saveRDS(metrics_df, file_out)
+    })
+  }
   start_date <- next_start_date
-}
-
-for (i in seq_len(nrow(date_ranges))) {
-  date_from <- date_ranges[i, "from"]
-  date_to <- date_ranges[i, "until"]
-  message(sprintf("{%s} - {%s}", date_from, date_to))
-  year_month <- gsub(pattern = "([[:digit:]]{4}-[[:digit:]]{2}).+", replacement = "\\1", x = date_from)
-  try({
-    metrics_df <- rsconnect::showMetrics(
-      metricSeries = "container_status",
-      metricNames = c("connect_count"),
-      appName = "ifnresource",
-      account = "rehwinkellab",
-      server="shinyapps.io",
-      from = as.numeric(as.POSIXct(date_from)),
-      until = as.numeric(as.POSIXct(date_to)),
-      interval = "60s"
-    )
-    file_out <- paste0(year_month, "-connect_count.rds")
-    saveRDS(metrics_df, file_out)
-  })
 }
